+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+- Register proposal ends soon event
+
 ## [0.0.28] - 2023-07-18
 
 ### Changed